+++ packages/happn-3/RELEASES.md
@@ -995,6 +995,10 @@
 -----------------
   - TEN-123: fix to getaddress - issue with bad breaking release in node 18.4
 
+13.1.4 2022-06-28
+-----------------
+  - TEN-12: mongo cluster membership
+