+++ packages/happner-2/RELEASES.md
@@ -969,6 +969,11 @@ SMC-1645: made error messages more consistent, encapsulating mesh.component.meth
 12.1.8 2022-06-28
 -----------------
   - TEN-123: fix to getaddress - issue with bad breaking release in node 18.4
+
+12.1.9 2022-06-28
+-----------------
+  - TEN-12: mongo cluster membership
+