+++ packages/happner-client/RELEASES.md
@@ -348,6 +348,9 @@
 -----------------
   - TEN-123: fix to getaddress - issue with bad breaking release in node 18.4
 
+12.2.8 2022-06-28
+-----------------
+  - TEN-12: mongo cluster membership