+++ packages/happner-cluster/RELEASES.md
@@ -385,6 +385,14 @@
 -----------------
   - TEN-123: fix to getaddress - issue with bad breaking release in node 18.4
 
+12.1.9 2022-06-28
+-----------------
+  - TEN-12: mongo cluster membership
+
+13.0.0 2022-06-28
+-----------------
+  - TEN-12: mongo cluster membership
+